+++ src/ast/passes/codegen_llvm.cpp
@@ -1842,29 +1842,6 @@ ScopedExpr CodegenLLVM::visit(Call &call)
     return ScopedExpr(buf, [this, buf]() { b_.CreateLifetimeEnd(buf); });
   } else if (call.func == "kstack" || call.func == "ustack") {
     return kstack_ustack(call.func, call.return_type.stack_type, call.loc);
-  } else if (call.func == "signal") {
-    bool target_thread = false;
-    if (call.vargs.size() == 2 &&
-        call.vargs.at(1).as<Identifier>()->ident == "current_tid")
-      target_thread = true;
-
-    auto &arg = call.vargs.at(0);
-    if (arg.type().IsStringTy()) {
-      auto signame = arg.as<String>()->value;
-      int sigid = signal_name_to_num(signame);
-      // Should be caught in semantic analyser
-      if (sigid < 1) {
-        LOG(BUG) << "Invalid signal ID for \"" << signame << "\"";
-      }
-      b_.CreateSignal(b_.getInt32(sigid), call.loc, target_thread);
-      return ScopedExpr();
-    }
-    auto scoped_arg = visit(arg);
-    Value *sig_number = b_.CreateIntCast(scoped_arg.value(),
-                                         b_.getInt32Ty(),
-                                         arg.type().IsSigned());
-    b_.CreateSignal(sig_number, call.loc, target_thread);
-    return ScopedExpr();
   } else if (call.func == "strncmp") {
     auto &left_arg = call.vargs.at(0);
     auto &right_arg = call.vargs.at(1);